fix: showing cache status in the table as well
